try to initalize store
